preprocessing/requirements.txt
astroid==2.4.2
attrs==20.1.0
aws-cdk.assets==1.61.1
aws-cdk.aws-applicationautoscaling==1.61.1
aws-cdk.aws-autoscaling-common==1.61.1
aws-cdk.aws-cloudformation==1.61.1
aws-cdk.aws-cloudwatch==1.61.1
aws-cdk.aws-codeguruprofiler==1.61.1
aws-cdk.aws-dynamodb==1.61.1
aws-cdk.aws-ec2==1.61.1
aws-cdk.aws-efs==1.61.1
aws-cdk.aws-events==1.61.1
aws-cdk.aws-iam==1.61.1
aws-cdk.aws-kms==1.61.1
aws-cdk.aws-lambda==1.61.1
aws-cdk.aws-logs==1.61.1
aws-cdk.aws-s3==1.61.1
aws-cdk.aws-s3-assets==1.61.1
aws-cdk.aws-sns==1.61.1
aws-cdk.aws-sqs==1.61.1
aws-cdk.aws-ssm==1.61.1
aws-cdk.aws-glue==1.61.1
aws-cdk.aws-s3-deployment==1.61.1
aws-cdk.cloud-assembly-schema==1.61.1
aws-cdk.core==1.61.1
aws-cdk.custom-resources==1.61.1
aws-cdk.cx-api==1.61.1
aws-cdk.region-info==1.61.1
cattrs==1.0.0
constructs==3.0.4
isort==5.4.2
jsii==1.12.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
publication==0.0.3
pylint==2.6.0
python-dateutil==2.8.1
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3
wrapt==1.12.1
boto3==1.14.56
